Fix/refactor based on suggestions

+++ src/main/java/com/emazon/stock/domain/usecase/CategoryUseCase.java
@@ -1,12 +1,13 @@
 package com.emazon.stock.domain.usecase;
 
 import com.emazon.stock.domain.api.ICategoryServicePort;
-import com.emazon.stock.domain.exception.DataConstraintViolationException;
-import com.emazon.stock.domain.exception.MissingAttributeException;
+import com.emazon.stock.domain.exception.CategoryAlreadyExistsException;
+import com.emazon.stock.domain.exception.CategoryNotFoundException;
 import com.emazon.stock.domain.model.Category;
 import com.emazon.stock.domain.model.PageCustom;
 import com.emazon.stock.domain.spi.ICategoryPersistencePort;
 import com.emazon.stock.domain.utils.PaginationValidator;
+import com.emazon.stock.utils.Constants;
 
 public class CategoryUseCase implements ICategoryServicePort {
 
@@ -18,35 +19,33 @@ public class CategoryUseCase implements ICategoryServicePort {
     
     @Override
     public void saveCategory(Category category) {
-        if (category.getName() == null || category.getName().isEmpty()) {
-            throw new MissingAttributeException("Category name cannot be empty");
-        }
-        if (category.getDescription() == null || category.getDescription().isEmpty()) {
-            throw new MissingAttributeException("Category description cannot be empty");
-        }
-        if(category.getName().length() > 50) {
-            throw new DataConstraintViolationException("Category name exceeds the maximum length of 50 characters");
-        }
-        if(category.getDescription().length() > 90) {
-            throw new DataConstraintViolationException("Category description exceeds the maximum length of 90 characters");
+
+        if(categoryPersistencePort.findByName(category.getName())){
+            throw new CategoryAlreadyExistsException();
         }
 
         this.categoryPersistencePort.saveCategory(category);
     }
 
     @Override
-    public PageCustom<Category> getAllCategories(int page, int size, String sortDirection, String sortBy) {
-        PaginationValidator.validatePagination(page,size,sortDirection,sortBy);
+    public PageCustom<Category> getAllCategories(Integer page, Integer size, String sortDirection, String sortBy) {
+        PaginationValidator.validatePagination(page,size,sortDirection);
         return this.categoryPersistencePort.getAllCategories(page,size,sortDirection,sortBy);
     }
 
     @Override
     public void updateCategory(Category category) {
+        if(!categoryPersistencePort.findByName(category.getName())){
+            throw new CategoryNotFoundException(Constants.EXCEPTION_CATEGORY_NOT_FOUND + category.getName());
+        }
         this.categoryPersistencePort.updateCategory(category);
     }
 
     @Override
     public void deleteCategory(String categoryName) {
+        if(!categoryPersistencePort.findByName(categoryName)){
+            throw new CategoryNotFoundException(Constants.EXCEPTION_CATEGORY_NOT_FOUND + categoryName);
+        }
         this.categoryPersistencePort.deleteCategory(categoryName);
     }
 }
